# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Capto mobile data-collection platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SearchParam"
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: language
          in: query
          schema:
            type: string
            enum: [en, hi]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryListResponse"
    post:
      operationId: createCategory
      tags: [categories]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
        - name: language
          in: query
          schema:
            type: string
            enum: [en, hi]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    patch:
      operationId: updateCategory
      tags: [categories]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
...
  /admin/faq:
    get:
      operationId: adminListFaq
      tags: [faq]
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: isActive
          schema: { type: boolean }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaqListResponse"
    post:
      operationId: adminCreateFaq
      tags: [faq]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFaqRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"
  /admin/faq/reorder:
    post:
      operationId: adminReorderFaq
      tags: [faq]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderFaqRequest"
      responses:
        "200":
          description: OK
  /admin/faq/{id}:
    get:
      operationId: adminGetFaq
      tags: [faq]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"
    patch:
      operationId: adminUpdateFaq
      tags: [faq]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFaqRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"
    delete:
      operationId: adminDeleteFaq
      tags: [faq]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: No Content

  # ─── Admin Support Settings ───────────────────────────────────────
...
  /admin/pages:
    get:
      operationId: adminListPages
      tags: [pages]
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaticPageListResponse"
    post:
      operationId: adminCreatePage
      tags: [pages]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaticPageRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaticPage"
  /admin/pages/{id}:
    get:
      operationId: adminGetPage
      tags: [pages]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaticPage"
    patch:
      operationId: adminUpdatePage
      tags: [pages]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaticPageRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaticPage"
    delete:
      operationId: adminDeletePage
      tags: [pages]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: No Content

  # ─── Admin Notices ────────────────────────────────────────────────
  /admin/notices:
    get:
      operationId: adminListNotices
      tags: [notices]
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: isActive
          schema: { type: boolean }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoticeListResponse"
    post:
      operationId: adminCreateNotice
      tags: [notices]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoticeRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notice"
  /admin/notices/{id}:
    get:
      operationId: adminGetNotice
      tags: [notices]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notice"
    patch:
      operationId: adminUpdateNotice
      tags: [notices]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoticeRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notice"
    delete:
      operationId: adminDeleteNotice
      tags: [notices]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: No Content

  # ─── Settings ─────────────────────────────────────────────────────
...
  /public/faq:
    get:
      operationId: getPublicFaq
      tags: [public-faq]
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: language
          schema: { type: string, enum: [en, hi] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Faq"
...
  /public/pages/{slug}:
    get:
      operationId: getPublicPage
      tags: [public-pages]
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
        - in: query
          name: language
          schema: { type: string, enum: [en, hi] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaticPage"
  /public/notices:
    get:
      operationId: getPublicNotices
      tags: [public-notices]
      parameters:
        - in: query
          name: language
          schema: { type: string, enum: [en, hi] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notice"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SearchParam:
      name: search
      in: query
      schema:
        type: string
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [total, page, limit, totalPages]
...
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        coverImageUrl:
          type: string
        displayOrder:
          type: integer
        isActive:
          type: boolean
        subcategoryCount:
          type: integer
        taskCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, isActive, displayOrder, subcategoryCount, taskCount, createdAt, updatedAt]
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        coverImageUrl:
          type: string
        displayOrder:
          type: integer
        isActive:
          type: boolean
      required: [name]
    UpdateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        coverImageUrl:
          type: string
        displayOrder:
          type: integer
        isActive:
          type: boolean
    CategoryListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, meta]
...
    Faq:
      type: object
      properties:
        id:
          type: string
        question:
          type: string
        answer:
          type: string
        displayOrder:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, question, answer, displayOrder, isActive, createdAt, updatedAt]
    FaqListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Faq"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, meta]
    CreateFaqRequest:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
        displayOrder:
          type: integer
        isActive:
          type: boolean
      required: [question, answer]
    UpdateFaqRequest:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
        displayOrder:
          type: integer
        isActive:
          type: boolean
    ReorderFaqItem:
      type: object
      properties:
        id:
          type: string
        displayOrder:
          type: integer
      required: [id, displayOrder]
    ReorderFaqRequest:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReorderFaqItem"
      required: [items]
...
    StaticPage:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        content:
          type: string
        isPublished:
          type: boolean
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, slug, content, isPublished, version, createdAt, updatedAt]
    StaticPageListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/StaticPage"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, meta]
    CreateStaticPageRequest:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        content:
          type: string
        isPublished:
          type: boolean
      required: [title, slug, content]
    UpdateStaticPageRequest:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        content:
          type: string
        isPublished:
          type: boolean
    Notice:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        isActive:
          type: boolean
        startsAt:
          type: string
          format: date-time
          nullable: true
        endsAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, title, content, isActive, createdAt, updatedAt]
    NoticeListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Notice"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, meta]
    CreateNoticeRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        isActive:
          type: boolean
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
      required: [title, content]
    UpdateNoticeRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        isActive:
          type: boolean
        startsAt:
          type: string
          format: date-time
          nullable: true
        endsAt:
          type: string
          format: date-time
          nullable: true
...